+++ docs/history.md
@@ -713,3 +713,4 @@ batch-matmul     1024      50        4.584        5.129      256.454     3747.70
 * tensor: strides, offset, select, slice
 * matmul for strides benchmark [qwen3 0.6 CUDA, TTFT=2129.780ms, Throughput: 0.516 tokens/s], 10x improvement
 * concat [benchmark no change]
+* added pytorch benchmark [qwen3 0.6 CUDA, TTFT=438.030ms, Throughput: 20.020 token/s]